requirements.txt
streamlit
tensorflow
torch
opencv-python-headless
Pillow
transformers
numpy
